fix: resolve all DTO and Mapper compilation errors

+++ src/main/java/ui/ft/ccit/faculty/transaksi/jenisbarang/view/JenisBarangService.java
@@ -1,7 +1,7 @@
-package ui.ft.ccit.faculty.transaksi.jenisbarang.view; // Sesuai struktur folder kamu
+package ui.ft.ccit.faculty.transaksi.jenisbarang.view;
 
-import ui.ft.ccit.faculty.transaksi.DataAlreadyExistsException; // Pastikan exception ini ada
-import ui.ft.ccit.faculty.transaksi.DataNotFoundException;      // Pastikan exception ini ada
+import ui.ft.ccit.faculty.transaksi.DataAlreadyExistsException;
+import ui.ft.ccit.faculty.transaksi.DataNotFoundException;
 import ui.ft.ccit.faculty.transaksi.jenisbarang.model.JenisBarang;
 import ui.ft.ccit.faculty.transaksi.jenisbarang.model.JenisBarangRepository;
 
@@ -24,8 +24,7 @@ public class JenisBarangService {
         return repository.findAll();
     }
 
-    // Ubah parameter id menjadi Byte sesuai Model
-    public JenisBarang getById(Byte id) {
+    public JenisBarang getById(Integer id) {
         return repository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("JenisBarang", id.toString()));
     }
@@ -36,7 +35,6 @@ public class JenisBarangService {
 
     // CREATE
     public JenisBarang save(JenisBarang jenisBarang) {
-        // Cek jika ID sudah di-set (biasanya untuk update, tapi kalau create id null/kosong)
         if (jenisBarang.getIdJenisBarang() != null && repository.existsById(jenisBarang.getIdJenisBarang())) {
             throw new DataAlreadyExistsException("JenisBarang", jenisBarang.getIdJenisBarang().toString());
         }
@@ -44,19 +42,14 @@ public class JenisBarangService {
     }
 
     // UPDATE
-    public JenisBarang update(Byte id, JenisBarang updated) {
-        JenisBarang existing = getById(id); // Akan throw DataNotFoundException jika tidak ada
-
-        // Update hanya field yang ada di Model
+    public JenisBarang update(Integer id, JenisBarang updated) {
+        JenisBarang existing = getById(id);
         existing.setNamaJenis(updated.getNamaJenis());
-
-        // Field lain (stok, harga, dll) DIHAPUS karena tidak ada di class JenisBarang
-
         return repository.save(existing);
     }
 
     // DELETE
-    public void delete(Byte id) {
+    public void delete(Integer id) {
         if (!repository.existsById(id)) {
             throw new DataNotFoundException("JenisBarang", id.toString());
         }
